Reorganized dbman
First, moved into the cppser directory.
Then we integrated the several modules into one file (dbman.h)
Finally removed the old files.
